Import of JESSE/Jifty-0.60505 from CPAN.
gitpan-cpan-distribution: Jifty
gitpan-cpan-version:      0.60505
gitpan-cpan-path:         JESSE/Jifty-0.60505.tar.gz
gitpan-cpan-author:       JESSE
gitpan-cpan-maturity:     released

+++ lib/Jifty/Test.pm
@@ -8,7 +8,6 @@ use Jifty::Server;
 use Jifty::Script::Schema;
 use Email::LocalDelivery;
 use Email::Folder;
-use Hook::LexWrap;
 
 =head2 import_extra
 
@@ -183,12 +182,14 @@ END {
         unlink mailbox();
 
         # Remove testing db
-        Jifty->handle->disconnect() if (Jifty->handle);
-        Log::Log4perl->get_logger("SchemaTool")->less_logging(3);
-        my $schema = Jifty::Script::Schema->new;
-        $schema->{drop_database} = 1;
-        $schema->run;
-        Log::Log4perl->get_logger("SchemaTool")->more_logging(3);
+        if (Jifty->handle) {
+            Jifty->handle->disconnect();
+            Log::Log4perl->get_logger("SchemaTool")->less_logging(3);
+            my $schema = Jifty::Script::Schema->new;
+            $schema->{drop_database} = 1;
+            $schema->run;
+            Log::Log4perl->get_logger("SchemaTool")->more_logging(3);
+        }
     }
 }
